Developments and refactors. Needs to be thoroughly tested.

+++ src/main/java/io/github/clamentos/cachecruncher/business/simulation/cache/CacheLine.java
@@ -14,7 +14,7 @@ public class CacheLine {
     ///
     private boolean valid;
     private boolean dirty;
-    private int tag;
+    private long tag;
 
     ///
 }
